Old-style domain literal warnings.
This mod is intended to help users transition from 1.5 domain literal syntax to
1.6.  The following warning will be generated when an array literal is encountered
which contains ranges:

"Encountered an array literal with range element(s). Did you mean a domain 
literal here? If so, use {...} instead of [...]. This warning can be disabled 
by using the --no-warn-domain-literal compiler option."

As the warning message indicates the new --no-warn-domain-literal compiler 
option has been provided which will allow users who actually desire to use
an array literal of ranges to silence the warning.

A handful of tests needed to be updated after this mod as the domain literal
syntax change did not change their semantic meaning.  A new test (warning.chpl)
has been added to validate the contents of the warning message.

NOTE:  This mod is only transitional!  It is intended that for releases after 1.6
this mod will no longer be needed and should be undone.  



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20264 3a8e244f-b0f2-452b-bcba-4c88e055c3ca