Parallel/distributed Edge Generation (EG) and Graph Construction (GC, aka Kernel 1) for SSCA2 in test/studies/ssca2/vass.
This is further development of the code in test/studies/ssca2/rmatalt,
ported to the current data structures of SSCA2 release version, as of r20245.
The old (release version) code is preserved (with the minimum of
necessary modifications) and can still be invoked by setting config
consts. The options are:

* Edge Generation
   default       : new code, in parallel
   --parEG=false : new code, some part of it running serially
                   (which part is a minor mod on the corresponding old code)
   --newEG=false : old code

* Graph Construction
   default       : new code
   --newGC=false : old code

* reproducibility
   default (with REPRODUCIBLE_PROBLEMS=true) :
      the resulting graph is reproducible, but two 'forall' loops in EG/GC
      are forced 'serial'
   -sSERIAL_GRAPH_GEN=false at compile time  :
      all parallelism is enabled, but the resulting graph will vary
      between executions
      (the edge histogram might still be reproducible)

For faster execution of Kernel 1, set these (*):
   SERIAL_GRAPH_GEN=false  (compile time)
   parGC=true              (default)

For faster overall execution, set these (*):
   SERIAL_GRAPH_GEN=false  (compile time)
   parGC=true              (default)
   newEG=true              (default)
   parEG=true              (default)
   edgeChecks=false        (skip consistency checks after EG)

(*) This gives up reproducibility. These settings are intended to give
better performance, but other settings may still be faster, e.g.,
on a single locale or for a serial execution.


Some notable changes:

* Edge-related arrays, including ones used in the preserved old code,
are now distributed using Block, by replacing edge_range with edge_domain.

* In the new edge generation (the default), random numbers are
generated using RandomPrivate_iterate() and used directly for
computation. Cf. the old code generated them using fillRandom() and
stored them in arrays.  The only intermediate array that the new edge
generation uses (if parEG) is permutation$.

* Collapsed multiple forall loops (mostly implicit in array computations)
into few explicit forall loops. The examples are the change to consistency
checks (see 'if edgeChecks') and the new vs. old noise computation
(see 'for depth in ...').

* Noted some TODO items in comments.

Additional changes:

* Added debugmsg() calls - do nothing now but can be hijacked for debugging.
Each call occurs *after* the code that it describes.

* Streamlined assign_quadrant(). Now it's a method on directed_vertex_pair,
modifying it directly, rather than through + on two directed_vertex_pairs.
The latter is now eliminated.

* Removed a couple of unused variables.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20266 3a8e244f-b0f2-452b-bcba-4c88e055c3ca