Add ref intents on proc arguments and use them to implement += type operators
Added TREF in the parser, handled the same as other intents. Changed += and
other op= operators to pass their full name to buildAssignment instead of
leaving the equals sign off.

Simplified buildAssignment to simply create a call the operator it is building.
Previously, it took a reference to the lhs and built 'lhs = lhs op rhs'.
Also, it had a special case for domains to create 'add' and 'remove' calls
for '+=' and '-='. The swap operator is also handled by buildAssignment now.

Added overloads of op= operators to ChapelBase.chpl to keep the current
semantics. In addition to a completely generic set, there are overloads for
sync types, iterator classes/records, and '+=' and '-=' for domains. For each
of these, there are overloads for param rhs and non-param rhs.

This change did not add overloads for ||= or &&=. These two operators need
special handling to avoid evaluating the second argument when it isn't needed.

The swap operator is also defined in ChapelBase for fully generic arguments,
plus one overload for sync lhs, sync rhs, and sync both sides. It is defined
in ChapelArray for arrays.

Create the reference type for intent ref args in the same place in
functionResolution where it is created for inout and out intents.  Add an
error for passing non-lvalues to ref intent args in the same place as for
out and inout.

When creating wrapper functions, the arguments to the wrapper need ref intent
when the wrapped function has ref args.

Add tests that use ref intents in a handful of different ways. There are tests
for iterators, class methods, regular procedures, recursive inner functions,
returning the reference as the return of a var function, and one that calls the swap operator on a bunch of different types of variables.

Changed test/functions/bradc/intents-error.good to reflect a slight change in
its expected error message.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20269 3a8e244f-b0f2-452b-bcba-4c88e055c3ca