Fix bug in pbs-aprun launcher
Changed the regular expression used to match the prompt to include the
newline in some cases.  This avoids the problem of the shell echoing
the typed command if sent when not at a prompt.

Remaining issues:

Periodically, it seems that qsub will read EOF from its pty, then end
the qsub interactive sessions before launching the program.  I don't
know why this is happening, and I've only seen it with moab/torque.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20273 3a8e244f-b0f2-452b-bcba-4c88e055c3ca