Added testing of the various SSCA2 edge generation and graph construction modes in test/studies/ssca2/vass.
Done as a separate test entry, rmatalt.chpl, which skips compilation
and execution of Kernels 2-4.

All modes use SERIAL_GRAPH_GEN=true in order for the resulting graph
to be deterministic. We could probably have somewhat meaningful
testing with SERIAL_GRAPH_GEN off as well, but I did not implement that.

Most of the testing was ported from test/studies/ssca2/rmatalt.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20275 3a8e244f-b0f2-452b-bcba-4c88e055c3ca