Minor change: Don't stop timers when clearing them
This commit makes it so that when .clear() is called on a timer,
the timer is not stopped.  This was the behavior previously,
though it's difficult to rationalize and was surprising to a
user.  With this change, the timer continues running if it was
running; remains stopped if it was stopped.  Updated the spec
to indicate this change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20282 3a8e244f-b0f2-452b-bcba-4c88e055c3ca