Removing pvm, mpi, and armci communication layers
Some time ago, we decided to remove these communication layers 
from the runtime on the argument that (a) they require work to
maintain, (b) we're not currently testing any of them, (c) they
cause noise when grepping for things, (d) resurrecting the code
from SVN and bringing it up to date later, if desired, is
probably less effort than keeping them alive when the runtime
interfaces change.

Finally getting around to actually doing it today.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20289 3a8e244f-b0f2-452b-bcba-4c88e055c3ca