This patch corrects a problem with inheritance in constructors reported by Rafael Larrosa.  An attempt to create an object in the derived class was failing because the default initializer called by every user-defined constructor in the derived class depends on the compiler-generated constructor in the base class. That constructor was being hidden when a user-defined constructor is present in the base class.
The approach used in this patch was to disable the code which hides the
compiler-generated constructor when any user-defined constructor is added to a
given class.  Then, the rules for disambiguation in function resolution were
modified, so that the compiler-generated (default) constructor is always
considered a weaker match that a user-defined constructor having the same
argument list.  

The expansion of generic constructors is greedy.  So making the default
constructor visible sometimes resulted in the default constructor shadowing the
user-defined constructor in generic expansion (which precedes disambiguation).
It was thus also necessary to add code to make the default constructor an
inferior choice during generic expansion.  This was done by splitting the loop
which traverses visible functions in gatherCandidates().  User-defined
candidates are considered first, and if any are found, they are used without
considering any compiler-defined candidates.

That change entailed two further changes:  It was necessary to remove the TEMP
flag from the compiler-supplied "main" function, so that it can override a
user-defined "main" in library code.  This is done so that the main in a library
module can be used as a hook for unit testing without intruding into the user's
namespace -- but only when that module is loaded through the module search path
(as opposed to being explicitly name on the command line).  It was necessary to
remove the TEMP flag from compiler-generated accessor functions.  The accessor
functions are really an internal device -- allowing field accesses to be treated
like a regular function call.  Since the field names *are* user-defined, it is
reasonable to give the accessor functions layered on top of them the same
binding precedence.


Detailed Comments:
-------- ---------
objRefAsFormArg.good
test_construct1.good
    Failing to set the INVISIBLE flag on the default constructor makes it
    visible in some cases where it was not before.  This is a change in the
    behavior of the compiler, but since it turns error cases into functioning
    cases, it is not a breaking change.  Looking at the source code for these
    test cases, it is not immediately obvious why they should work.  Subject to
    team approval, the corresponding section of the specification should be
    updated prior to checkin, so that the observed behavior is correct relative
    to the spec.

base-initializer.chpl
base-initializer.good
    Tests added to lock in the new behavior.

base-class-generic-args.chpl
base-class-generic-args.future
base-class-generic-args.good
	Rafael found another bug in constructor inheritance.

brad-dispatch.chpl
brad-dispatch.future
brad-dispatch.good
    Brad supplied a test for correct binding.
    The binding part works, there is still a problem with dynamic dispatch.

Modules.tex
	Expanded slightly on how when a default "main" function is generated, and
	what it looks like.

normalize.cpp:797,811-72
    Removed a redundant conditional and reindented.
normalize.cpp:1137-9
    This is where the INVISIBLE flag is no longer set.

buildDefaultFunctions.cpp:157
    Removed the TEMP flag from accessor functions, so they have the same binding
    precedence as the fields they access.
buildDefaultFunctions.cpp:287
    Removed the TEMP flag from the compiler-generated main function, so it is
    considered in preference to main functions that are defined in library
    modules.

functionResolution.cpp:1457-69
    Special test added to disambiguateByMatch(), making the default constructor
    inferior to any user-defined constructor.
functionResolution.cpp:2159-99,2256-73
    In gatherCandidates(), split the loop over visibleFns so that user-defined
    functions are considered first, and compiler-generated functions are
    considere only if no user-defined candidates are found.  Also, moved the
    --explain-call code out to the parent routine.
functionResolution.cpp:4543-72,
    Factored buildValueFunction() out of resolveFns(), and simplified it by
    consolidating conditional clauses having the same condition.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20291 3a8e244f-b0f2-452b-bcba-4c88e055c3ca