This commit reorders the execution options that sub_test applies
[reviewed by Sung]

Prior to this commit, execution options were supplied by sub_test for
a test foo.chpl in the order:

    './EXECOPTS' '--execopts' 'foo.execopts'

This commit reorders the flags as follows:

    './EXECOPTS' 'foo.execopts' '--execopts'

on the argument that if you're applying a set of execution options via
a globally-set flag that's not baked into the test system, you
probably want it applied last (if there's a good order at all to
always supply it).  E.g., if you want to compile everything -O0, you
wouldn't want a local -O3 to the test to override it.

The current rationale for this change is that some tests print out an
error in their argument lists that's numbered by which argument
contained the error, and putting a global execopts like this first
messes things up.  In qthreads testing, for example, three tests are
currently failing because we run qthreads gasnet testing with
--numThreadsPerLocale in order to avoid oversubscription.

Unfortunately, this test only fixes 2 of the 3 failures because the
third checks for an error when the last argument is missing its
value but now that we've supplied an additional argument, the error
comes out differently.  Probably the right thing to do here is to
permit the number of threads per locale to be set by an environment
variable; or more generally to permit any config to be set via an
environment variable.  But for now, 2 out of 3 ain't bad.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20299 3a8e244f-b0f2-452b-bcba-4c88e055c3ca