Committing 2 futures related to remote value forwarding
These futures are motivated by the RA regression last night (caused by
sync vars disabling the remote value forwarding of constant stuff like
array descriptor identity) and recent observations by Vass and Sung
that certain things aren't getting remote value forwarded in spite of
their const-ness in the presence of sync vars.

The full test I wanted to commit is remoteValueForwardArrayDescsConsts
which tests a variety of idioms that should all be safe w.r.t. the
local block, but fail due to the sync vars presence and the fact that
they're disabling the remote value forwarding of array desciptors
and constants.

I have a patch that fixes the array descriptors case, undergoing
testing and review by Sung now.

The const case is a little more complicated as noted in the .future
(along with my thoughts about how to fix/improve).  I may take that on
next when I find the time.

Sung and I have recently been talking about what it would take to
retire the local block while still getting the benefits that it
provides.  One idea is to pursue the notion of local variables more
aggressively and have a means of capturing a local sub-array that
permits communication-free access to a local array's elements (view
this as a wrapper around the distribution's local array descriptor).
It seems that such a concept would give us most of the benefit we
get from local blocks today while removing the confusion of applying
local to a whole block of code and tying it to specific variables
instead.  This also overlaps with user requests to have a means of
getting access to "the current locale's block of data" and the like.
But that's a bigger change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20304 3a8e244f-b0f2-452b-bcba-4c88e055c3ca