Improved remote value forwarding for array descriptors
[reviewed by Sung; tested on multilocale tests, but will likely result
in failures due to improved comm counts for some of the distribution
robustness tests and other directories that we don't typically require
testing for multilocale testing -- I'll clean up tomorrow.]

This commit improves remote value forwarding for array descriptors,
and sets the stage for improving remote value forwarding of other
constant/invariant data.

Prior to this change, the presence of a sync var within a function
would disable remote value forwarding for all arguments to that
function as a means to enforce the Chapel memory consistency model.
However, constant data should still be permitted to be remote value
forwarded since it can't change.  Vass pointed out cases of this
recently and Sung may have run into them not so long before.  The
motivation for this change came from Greg's RA fix yesterday which
failed on multiple locales due to the newly-added sync var accesses
disabling remote value forwarding for the table T's privatized array
descriptor.

For arrays, whether variable or constant, their array descriptors are
invariant and should be permitted to be remote value forwarded.  For
non-privatized arrays, you can't really remote value forward the array
descriptor because it's a pointer to a class on a single locale; but
for privatized arrays, the pointer is replaced with a unique ID
(stored as an int) that is amenable to remote value forwarding.

The main part of this change is to push the test for sync var
references further into the logic of the remote value forwarding
analysis to serve as a reason to bail out of considering an argument
rather than as the first step to avoid consideration of a function at
all.  A new routine isSufficientlyConst() determines whether an
argument is sufficiently constant/invariant to be considered for
remote value forwarding even in the face of sync var accesses.
Currently, it only returns true for privatized array types (note that
the "privatized" part of this fell out naturally -- due to the way
non-privatized array types are represented in the IR, they did not
pass the test).

I replaced the old conditional with an if (true) for the time being to
minimize the diff in the SVN logs; I'll check in a beautifying change
right after this to remove that and futz with all the indenting.

I had hoped that this would lead to being able to remote value forward
scalar constants as well, but that is a bit more challenging due to
their being placed on the heap and passed through a few layers of
routines prior to being considered -- this is true for the array
descriptors as well of course, but their types uniquely identify them
as being candidates, whereas an 'int' type could be either const or
not, so we need to look deeper.  My proposed fix for this case would
be to actually remove const scalar types from being placed on the
heap as other variables are, which would essentially force them to
be remote value forwarded (I think).  Didn't have time to try that
today, and may not for awhile yet.

Retired the future I checked in this afternoon and updated some comm
counts in the robustness suite due to not needing to chase after
array descriptors.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20305 3a8e244f-b0f2-452b-bcba-4c88e055c3ca