Disabling the removeEmptyRecords pass by default
When QIO was committed, the removeEmptyRecords pass was disabled
by having it simply 'return' for reasons unknown at the time.
The rationale was that the importance of QIO was higher than
that of the removeEmptyRecords pass.

This week, as part of Triage, I went in to try and figure out
why QIO caused this failure, as well as to determine what the
value of this pass is/was.  As a result of this study, I
decided that it was reasonable for the pass to be disabled for
now, but disabled it by reversing the default value for the
existing developer flag rather than hard-coding a return into
the pass, as was done in the QIO change.  This permits us to
turn it on for a given compile if, at some point in the future,
we want to explore re-enabling it.

The analysis for why QIO broke this pass is that it uses empty
records with methods attached to them (e.g., to represent a
newline).  The pass was trying to analyze an unresolved call 
that was taking one of these empty records as an argument --
I did not get far enough to understand why it was unresolved;
I did try simply skipping over such unresolved calls, but this
broke the correct behavior (newlines were essentially dropped
on the floor).  It could be that the right thing to do would
be to remove empty records only if they have no methods attached
to them, but this raises the question of why such empty records
might exist in the first place.  Alternatively, we could dig
deeper to find out why this call was unresolved.

Instead, I went back to try and understand why the pass was
inserted in the first place.  Steve's commit message said
that prior to the pass, we were getting C compiler warnings
when compiling ra.chpl with -O and --devel, but going back
to release versions of the compiler prior to his change and
then the SVN version of the compiler just prior to his
change, I was unable to reproduce it.  Best guess is that
a difference in gcc versions caused different behavior.

Based on being unable to come up with a case where this was
clearly beneficial, keeping it disabled by default seemed
like a reasonable thing to do.

It occurs to me as I type this that I should file a future
capturing an instance of the empty record case with the flag
on, just to capture that issue.  Will do that next.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20310 3a8e244f-b0f2-452b-bcba-4c88e055c3ca