Fix bug in error handling of fork'ed utility program in launcher
[reviewed by gbt]

This fix gives a better error message when a forked process fails with
error in chpl_run_utility1K().  This function is used to run short
programs from which we want to capture the output (e.g., qsub
--version).  In the child, if the execvp() fails, the parent has no
way of knowing except by examining the output.  Added a search for the
string "internal error:" in the output to catch this case.  The string
is reasonably unique, but not ideal.  If this turns out to be an
issues, the we can make the Chapel runtime internal error string more
interesting (maybe put the PID in there).

This was motivated by annoying user experience: Brad and Rachel both
ran into the weird error message when they had not loaded the pbs
module.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20312 3a8e244f-b0f2-452b-bcba-4c88e055c3ca