Removing svn checkout of third-party-norelease directory
We used to use this for testing experimental code, but haven't
relied on it for some time, so are removing it to avoid the
time/disk overhead associated with it.  People who run nightly
tests may want to clean up their chpl-third-party-cron-*-readonly/
directories to save space.

In the future, if we re-enable this, we may want to do it conditionally
for only those test configurations that require the directory.  For
now I just removed it to keep things simple and clean.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20321 3a8e244f-b0f2-452b-bcba-4c88e055c3ca