Remove script's dependence on CHPL_HOME
Prior to this change, this script required CHPL_HOME to be set and
would update the SVNLOG file in CHPL_HOME.  As of this change, the
script can be run in any svn directory and will generate a log for
that directory.  This mean, if you relied on the old behavior you
will want to cd $CHPL_HOME before running the script.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20322 3a8e244f-b0f2-452b-bcba-4c88e055c3ca