Provide the total number of edges in the Graph data structure
[reviewed by gbt]

Provide the total number of edges in the Graph via num_edges.

For the RMAT case in kernel 1, we either compute the total number of
edges in the graph or read it in from a file, so there is no need to
recompute it in kernel 4 for the TEPS computation.

For the torus cases, we do not compute the total number of edges, so I
just provided a parenthesis-less function that does the reduction.
I'm sure there's a formula for this, but the current data structures
are not set up to put in a generic computation for this, so I opted
for the minimal fix.

Changed kernel 4 to use num_edges, and also changed a some vars to
consts since I was in the area.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20327 3a8e244f-b0f2-452b-bcba-4c88e055c3ca