Added lazy initiation of the RAD cache to Cyclic and Block distributions. [Reviewed by Sung.]
This feature is toggled on/off by the config params disableBlockLazyRAD and
disableCyclicLazyRAD (lazy initialization is on by default, as configured in 
DefaultRectangular by the value defaultDisableLazyRADOpt). When initialization
is lazy, the array's RAD cache is only allocated when it is first needed (ie,
on the first access to a non-local element). Furthermore, the remote locale's
array access info for each locale is fetched on the first access to that locale,
rather than at array allocation time.  If the array is resized or reindex, the
existing RAD cache will be deleted.

I also found and removed a number of unnecessary gets in the non-lazy case; after
this commit the performance of non-lazy RAD cache initialization is also better.

To maintain correct behavior, I also added a lock to LocBlockArr class and 
lock/unlock functions for allocating the RAD cache.

(Note: although we considered making the RAD configuration an enum, this 
does not allow for on-clause optimization even in the case where lazy 
initialization is turned off because the arrays are not set up to consider 
this value at compile time.)

This optimization improves the overall runtime of SSCA2 on large numbers of
locales (300+) by 3x. :)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20328 3a8e244f-b0f2-452b-bcba-4c88e055c3ca