Modify verification so as to not introduce additional errors.
[This duplicates the change of r20301 into this version of RA.  The
original r20301 change was reviewed by Brad.]

Add an array of sync variables over a strided domain corresponding to
that of the main data table T.  Use each element of the sync var array
as a lock that single-threads access to a stride-sized block (referred
to as a "protection block") of elements in the main table.  Provide
configuration constants that parameterize the stride and thus the size
of the "protection block", with a default chosen to balance memory use
against conflict avoidance overhead.

While here, I also made minor corrections to a couple of comments that
were no longer accurate because they weren't updated as the code was
changed.

All of this is done to prevent false failures due to errors introduced
during the verification step itself, rather than introduced during the
main computation and discovered during verification.  In nightly
testing, RA has been fluctuating fairly between passing and failing,
and this change should reduce those fluctuations.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20333 3a8e244f-b0f2-452b-bcba-4c88e055c3ca