Updates to aprun and pbs-aprun launchers
[reviewed by gbt]

This commit includes a few updates to the aprun and pbs-aprun launchers:

- Add -j option to aprun if supported by ALPS (>5.X)

- Add flag --generate-qsub-script that writes out a basic qsub script
  for running the Chapel program when using the pbs-aprun launcher

The commit also starts the refactoring of common code across launchers
with common parts (e.g., aprun and pbs-aprun).  The file aprun-utils.c
contains the code that determines what arguments to pass to aprun.  It
is currently checked in under aprun and pbs-aprun, but we expect to come
up with a better way to organize this.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20336 3a8e244f-b0f2-452b-bcba-4c88e055c3ca