Change the atomics used in DefaultAssociativeDom into explicit processor atomics
[reviewed by gbt]

This commit makes the atomics used for numEntries and tableLock in
DefaultAssociativeDom explicit processor atomics.  DefaultAssociative
is a single locale domain map, and, at least currently, it is used
primarily from the locale on which it allocated.  Note that by design,
the algorithms used are cognizant of the fact that processor atomics
are being used and there are explicit on clauses sprinkled throughout
the code.  If I were doing this with network AMOs, there's a good
chance I'd want to do something a bit different.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20341 3a8e244f-b0f2-452b-bcba-4c88e055c3ca