This patch adds the ability for the compiler to infer CHPL_HOME sometimes.
To decide if chpl is run from a CHPL_HOME directory,
we check for exepath/../../util/chplenv. If that exists, we
infer CHPL_HOME to be exepath/../..

If we have an inferred CHPL_HOME and it does not match the environment
CHPL_HOME, we print a warning. But, in all cases the environment
variable takes precedence. Hopefully this patch will reduce
confusion when working with multiple versions of the compiler.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20348 3a8e244f-b0f2-452b-bcba-4c88e055c3ca