Standardize the task side of the cooperative comm/task interface.
We were allowing the tasking layer side of the cooperative comm/task
interface to be defined however the tasking layer wanted, by means of a
preprocessor macro.  Replace this with a standard interface function.
The tasking layer can still either provide it or not, but if it does
provide it, it has to match the interface for chpl_comm_get().


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20351 3a8e244f-b0f2-452b-bcba-4c88e055c3ca