Added a script to save away the current test logs.
Every time it runs, it stashes away each log file into NightlyHistory/yymmdd,
where yymmdd is the "Starting Chapel regression tests" date listed in the log.
Command-line arguments specify the directories where to look for the logs.
The log files are the files named day*.log .
If the log has not changed since it was stashed last, copying is not performed.

I use 'rsync -t' for updated-only copying, rather than 'cp -up'. The
latter does unneeded copies, here's why, according to my experiments.
The time stamp of the rsync or cp destination drops the fractional seconds
of the source, making the destination younger than the source.
When run the second time around, assuming that the source has not changed,
'cp -u' copies the file again, whereas 'rsync -t' doesn't.
We want the latter.

The ideal solution, though, would be to augment 'nightly' itself to stash
away a copy of a log once it's ready. In which case the destination
directory name could also have the revision number in it, besides the date.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20369 3a8e244f-b0f2-452b-bcba-4c88e055c3ca