Limit number of threads for valgrind testing
Limit the number of threads (--numThreadsPerLocale) for valgrind
testing.  This seems to help with the current remaining valgrind "the
'impossible' has happened" errors.  Specifically, all the 'impossible'
errors from last weekend's testing were due to "VG_N_THREADS is too
low".  Since we don't want to rebuild valgrind for every system we
have, I have limited the number of threads as we do for qthreads
testing.  In hand testing, this go rid of all the 'impossible' errors.
Hopefully it will work in the nightly.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20391 3a8e244f-b0f2-452b-bcba-4c88e055c3ca