Added floating point atomics (intrinsics and locks).
[Reviewed by Sung]
Added processor atomic types for real(32)s and real(64)s in both intrinsics
and locks.  and, xor, and or are not defined on real atomics for obvious reasons
(although they are defined on integer type atomics).

The intrinsic version relies on a double being 64 bits and a float 
32 bits.  Since chpltypes in general require this to be the case, this is
not an extra restriction.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20392 3a8e244f-b0f2-452b-bcba-4c88e055c3ca