setchplenv scripts remove old PATH and MANPATH using CHPL_HOME
That way, a user can switch Chapel projects without logging out and logging
back in, or when they have setchplenv source'd in their shell startup files.

I did it with a short Perl script, because I didn't want to figure out how to
do some regexping in each of the 4 shells.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20393 3a8e244f-b0f2-452b-bcba-4c88e055c3ca