Updates to REGRESSIONS and TRIAGE files
Took on valgrind testing during my TRIAGE week.  Good news it that
there was only one type of 'impossible has happened' error.  Bad news
is that the suggested fix is to recompile valgrind with a VG_N_THREADS
> 100.  So instead, I limited the numThreadsPerLocale to 100 for all
tests (similar to what we do for qthreads testing).  That seemed to
help in hand testing.

More bad news.  There are two (new) real valgrind errors.  One is in
the generated code when calling _build_range(), the other is in the
compiler when compiling the primer chpldoc.chpl, though it has nothing
to do with chpldoc (IOU a simple test case future).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20396 3a8e244f-b0f2-452b-bcba-4c88e055c3ca