Fixes bugs with QIO writebits
[reviewed by Tom]

Especially we were seeing problems when writing a lot of 64-bit numbers with
writebits. Besides fixing two bugs with bit-I/O, this patch adds additional
tests.

One problem was just a type-o, the other was that shifts >= word size do not
result in zero, and so in some cases we have to arrange for zero to be the
result of the shift...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20400 3a8e244f-b0f2-452b-bcba-4c88e055c3ca