Update PREDIFF script to remove "configured to" string for built-in config vars
This commit updates the PREDIFF script for these tests to remove any
output for configured built-in config vars.  As a result, I also
removed the '-nl 1' in all the execopts files that was there to ensure
so that the result for CHPL_COMM=none was same as gasnet.

This was motivated by the update to use the global execopts
--numThreadsPerLocale for valgrind testing (r20391), which resulted in
a number of new failures due to expected help output (oops).  Note,
this would also happen for qthreads testing if we tested the entire
suite.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20404 3a8e244f-b0f2-452b-bcba-4c88e055c3ca