Promote changes from r20259, r20315, r20360, r20403 to the release version of SSCA2
- Limit total number of task private variables (r20315)
- Add use of network atomics (avoid on clauses when using them) (r20259)
- Add use of floating point atomics even in the processor atomics case (r20403)
- promote numEdges change (r20360)

This commit also updates the comm diags test prediff and good file.
Specifically, due to the limited number of task private variables, the
number of forks is slightly higher and the number of gets is slightly
more variable.  For the gest, the best case is when the acquired task
private variable is local (resulting in same value as before), the
worse case is less that 10% greater than the best case.

At some point it seems that the number of forks stopped being
variable, so reflected that change in the prediff script.

Update the README for the comm diags test to describe the change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20408 3a8e244f-b0f2-452b-bcba-4c88e055c3ca