Added two flags for altering the documentation output to suit the user's preferences.
--docs-folder <foldername> changes the output folder name from docs/ to 
<foldername>.  This change is encapsulated in compiler/include/driver.h, 
compiler/main/driver.cpp, and compiler/passes/docs.cpp, with an update to 
test/compflags/bradc/help/userhelp.good and the addition of 
test/chpldoc/compflags/folder/* to test its functionality.

--docs-comment-style <indicator> limits the output into the documentation 
to only those comments which start with the provided indicator.  If the 
indicator itself does not start with '/*', then an error message stating 
the problem is displayed.  The comments which start with the provided 
indicator are assumed to terminate with a reflection of this indicator.  
This change is encapsulated in compiler/include/driver.h, 
compiler/main/driver.cpp, and compiler/parser/processTokens.cpp, with an 
update to test/compflags/bradc/help/userhelp.good and the addition of 
test/chpldoc/compflags/comment/* to test its functionality.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20410 3a8e244f-b0f2-452b-bcba-4c88e055c3ca