dtLocale implementation: Replace the representation of a locale ID by "int" with a struct that includes "node" and "subloc" fields.
This is the first half of a full implementation: The node ID is now written into
and read out of the structure, but the subloc field is as yet unused.
Beyond the simple rewriting, the role of the node ID (== GASNet node ID) was
clarified.  Two assumptions are represented by this change:
 1. The memory address space on each "node" is flat -- in the sense that memory
 allocated in one sublocale can be distinguished from memory allocated in a
 different sublocale on the same node solely on the basis of the .addr portion
 of their respective wide pointers.
 2. Communication is not required when one sublocale on the same node accesses
 data allocated by a different sublocale on the same node.


NOTES
- The "as_int" representation of a locale ID is a convenience which can probably
be removed at a later date.  The union can then be removed, initializers
simplified, etc.  It was used because it allows a default locale ID to be
supplied as an int(64) literal, which is just a bit simpler than declaring a
temp, initializing the temp to {0,0}, and then passing the temp in as a record.
- At present, Chapel is unaware of the structure of a locale ID.  Parsing out
the fields is all handled in the code generator, or through primitives.  It
probably makes sense to expose these fields in the language and remove the
primitives.


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/caseyb@20412 3a8e244f-b0f2-452b-bcba-4c88e055c3ca