The partial implementation of the Chapel Meteor solution.
Within Meteor.chpl are implemented several classes required for the solution 
as outlined within a Java solution linked within the README: Cell, PieceCell,
portions of Piece and a new class (PieceList) created to encapsulate the cells
in a useful fashion.  The comments for these classes are extensive, and 
hopefully comprehensive.  BoardCell and Board classes are created as stubs
but entirely unimplemented.  The algorithm itself is unimplemented.

There are a number of tests in the directory "tests", none of which are in
start_test form.  They do give extensive output relative to the actions of the
implemented classes after which they are named, which should be useful to those
seeking to determine the correctness of the classes implemented without 
converting the tests to a form recognizable by start_test.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20420 3a8e244f-b0f2-452b-bcba-4c88e055c3ca