Created a chpldoc mode for generating creole tags that can be used by a script I also wrote to convert the tagged text files into HTML files.
docs.cpp has been altered to generate the creole markup when the --docs-no-creole flag is not set.  driver.h and driver.cpp have been altered to create this flag.  The majority of the effort for this portion lies in the util/docs directory.  Here lies a script, chpldoc2html, which takes either a provided .txt file or a directory, and turns the .txt files into .html files.  official.css contains the style that is automatically used by this script, and plain.css contains a simplier version.  Full documentation for the script lies within the script itself, sorry I only have 5 minutes left to commit!

Created tests and altered old ones to reflect the new changes.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20422 3a8e244f-b0f2-452b-bcba-4c88e055c3ca